+++ CHANGELOG.md
@@ -9,10 +9,12 @@
 - Added support for query `PeerList`
 - Added support for query `PeerStats`
 - Added support for query `PeerVersion`
+- Added support for query `GetAncestors`
 - Added support for query `NodeInfo`
 - Added support for query `JoinNetwork`
 - Added support for query `LeaveNetwork`
 - Added support for query `GetInstances`
+- Added support for query `GetBranches`
 
 ## 3.0.0
 - Stronger typing for various places in the API.